Fix index overflow in backward db2/ds kernel: int64 token/scatter indices
Completes #23 for the backward path. #52 fixed the forward gather
(reduction_over_k_gather.py) by upcasting its index math to int64, but
_compute_db2_and_ds_kernel still loads token_indices and scatter_indices
as tl.uint32. token_indices * H (the dout gather) wraps once T * H >=
2**32 — T >= 1,048,576 tokens at H = 4096, or 524,288 at H = 8192 —
silently gathering dout rows from wrapped in-bounds addresses, i.e.
corrupt db2/ds gradients with no error raised. scatter_indices spans
Mtotal = T * K and multiplies by OLD_DS_PARTIAL_N and NUM_H_BLOCKS;
further from the edge, but int64 costs nothing in this memory-bound
kernel.

Found while root-causing a train/inference divergence in a 120B hybrid
MoE (K=22): the forward instance of this overflow class corrupted ~27%
of expert contributions at 262,144-token packed batches on v0.1.2.post1
(pre-#52), bit-exact reproduction at <= 180,224 tokens and monotone
corruption growth from 196,608 up, matching 2**32/(K*stride) exactly.

+++ sonicmoe/functional/backward.py
@@ -69,10 +69,10 @@ def db2_and_ds_kernel(
         tk_grouped = E_count_start + tk_offsets
 
         # Gather token indices: [BLOCK_TK]
-        token_indices = tl.load(x_gather_idx_ptr + tk_grouped, mask=tk_mask, other=0).to(tl.uint32)
+        token_indices = tl.load(x_gather_idx_ptr + tk_grouped, mask=tk_mask, other=0).to(tl.int64)
 
         # Get scatter indices: [BLOCK_TK]
-        scatter_indices = tl.load(s_scatter_idx_ptr + tk_grouped, mask=tk_mask, other=0).to(tl.uint32)
+        scatter_indices = tl.load(s_scatter_idx_ptr + tk_grouped, mask=tk_mask, other=0).to(tl.int64)
 
         s = tl.load(s_ptr + scatter_indices, mask=tk_mask, other=0.0).to(tl.float32)
